+++ vdirsyncer/storage/dav.py
@@ -719,6 +719,8 @@ class CalDAVStorage(DAVStorage):
     _property_table.update(
         {
             "color": ("calendar-color", "http://apple.com/ns/ical/"),
+            "description": ("calendar-description", "urn:ietf:params:xml:ns:caldav"),
+            "order": ("calendar-order", "http://apple.com/ns/ical/"),
         }
     )
 
@@ -835,3 +837,13 @@ class CardDAVStorage(DAVStorage):
             </C:addressbook-multiget>"""
 
     get_multi_data_query = "{urn:ietf:params:xml:ns:carddav}address-data"
+
+    _property_table = dict(DAVStorage._property_table)
+    _property_table.update(
+        {
+            "description": (
+                "addressbook-description",
+                "urn:ietf:params:xml:ns:carddav",
+            ),
+        }
+    )